monitoreo-e-ingesta-de-datos/requirements.txt
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyspark==4.1.1
